+++ tests/general/api/SupportedDataTypes.cpp
@@ -1,7 +1,7 @@
 /*
  -------------------------------------------------------------------------------
     This file is part of the test suite for weather-information-collector.
-    Copyright (C) 2021, 2022  Dirk Stolle
+    Copyright (C) 2021, 2022, 2025  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -78,7 +78,7 @@ TEST_CASE("Check supported data types of API implementations")
 
   SECTION("Weatherbit")
   {
-    Weatherbit wb(PlanWeatherbit::Starter, "This is not a valid key.");
+    Weatherbit wb(PlanWeatherbit::Standard, "This is not a valid key.");
     // Weatherbit supports current and forecast data.
     REQUIRE( wb.supportsDataType(DataType::Current) );
     REQUIRE( wb.supportsDataType(DataType::Forecast) );