+++ changelog.md
@@ -5,8 +5,12 @@
 The request limit for Weatherstacks's free plan is adjusted to 100 requests per
 month, since they changed the limit from 1000 to 100 requests.
 
-Support for Weatherbit's Hobbyist plan is removed, because Weatherbit does not
-offer that plan anymore.
+There have been a few changes in Weatherbit's plans, and this release reflects
+those changes:
+* Support for Weatherbit's Hobbyist plan is removed, because Weatherbit does not
+  offer that plan anymore.
+* The Starter plan is renamed to Standard and the limit is changed from 50000
+  requests per day to 25000 requests per day.
 
 simdjson, the library used for JSON parsing, has been updated from version
 3.11.1 to version 3.11.6.